+++ source/be_runtime.h
@@ -6,7 +6,7 @@
 #include "reswords.h"
 
 void de_reference(s1_ptr a);
-void UserCleanup(int status);
+void UserCleanup(long status);
 
 #define FIRST_USER_FILE 3
 #define MAX_USER_FILE 40
@@ -14,19 +14,19 @@ extern struct file_info user_file[MAX_USER_FILE];
 
 extern struct routine_list *rt00;
 
-extern int EuConsole;
+extern long EuConsole;
 
 extern unsigned char *string_ptr;
 extern char *file_name_entered;
-extern int warning_count;
+extern long warning_count;
 extern char **warning_list;
-extern int crash_count;
-extern int clocks_per_sec;
-extern int clk_tck;
-extern int gameover;
-extern int insert_pos;
+extern long crash_count;
+extern long clocks_per_sec;
+extern long clk_tck;
+extern long gameover;
+extern long insert_pos;
 
-extern int TraceOn;
+extern long TraceOn;
 extern object *rhs_slice_target;
 extern s1_ptr *assign_slice_seq;
 
@@ -34,19 +34,19 @@ extern object last_w_file_no;
 extern IFILE last_w_file_ptr;
 extern object last_r_file_no;
 extern IFILE last_r_file_ptr;
-extern int current_screen;
+extern long current_screen;
 
 extern IFILE TempErrFile;
 extern char *TempErrName; // "ex.err" - but must be on the heap
 extern char *TempWarningName;
-extern int display_warnings;
+extern long display_warnings;
 
 extern long seed1, seed2;  /* current value of first and second random generators */
-extern int rand_was_set;
-extern int con_was_opened; /* TRUE if CON device was ever opened */
-extern int current_screen;
+extern long rand_was_set;
+extern long con_was_opened; /* TRUE if CON device was ever opened */
+extern long current_screen;
 
-extern int print_chars;  // value can be checked by caller
+extern long print_chars;  // value can be checked by caller
 
 extern struct op_info optable[MAX_OPCODE+1];
 
@@ -86,51 +86,51 @@ int key_write;;       // place where next key will be stored
 void system_call(object command, object wait);
 object system_exec_call(object command, object wait);
 
-int might_go_screen(object file_no);
-void set_text_color(int c);
+long might_go_screen(object file_no);
+void set_text_color(long c);
 
 void Append(object_ptr target, object s1, object a);
 void Prepend(object_ptr target, object s1, object a);
 void Replace( replace_ptr rb );
 void Concat(object_ptr target, object a_obj, object b_obj);
 s1_ptr Add_internal_space(object a,int at,int len);
-void Concat_Ni(object_ptr target, object_ptr *source, int n);
+void Concat_Ni(object_ptr target, object_ptr *source, long n);
 
 object EGetEnv(object name);
 
-object EPrintf(int file_no, object format_obj, object values);
-void StdPrint(int fn, object a, int new_lines);
+object EPrintf(long file_no, object format_obj, object values);
+void StdPrint(long fn, object a, long new_lines);
 void EPuts(object file_no, object obj);
-void Print(IFILE f, object a, int lines, int width, int init_chars, int pretty);
-int show_ascii_char(IFILE print_file, int iv);
+void Print(IFILE f, object a, long lines, long width, long init_chars, long pretty);
+int show_ascii_char(IFILE print_file, long iv);
 
-int get_key(int wait);
+int get_key(long wait);
 object EGets(object file_no);
 void EClose(object a);
 int CheckFileNumber(object a);
 int NumberOpen();
 void key_gets(char *input_string);
 
-IFILE which_file(object a, int mode);
-void MakeCString(char *s, object pobj, int slen);
+IFILE which_file(object a, long mode);
+void MakeCString(char *s, object pobj, long slen);
 
 void setran();
 void call_crash_routines();
 
-int compare(object a, object b);
+long compare(object a, object b);
 object calc_hash(object a, object b);
 void ctrace(char *line);
 void Position(object line, object col);
-extern int charcopy(char *, int, char *, int);
-s1_ptr Copy_elements(int start,s1_ptr source, int replace );
+extern long charcopy(char *, long, char *, long);
+s1_ptr Copy_elements(long start,s1_ptr source, long replace );
 cleanup_ptr ChainDeleteRoutine( cleanup_ptr old, cleanup_ptr prev );
-cleanup_ptr DeleteRoutine( int e_index );
+cleanup_ptr DeleteRoutine( long e_index );
 void AssignSlice(object start, object end, s1_ptr val);
 void cleanup_double( d_ptr dbl );
 void cleanup_sequence( s1_ptr seq );
-void Tail(s1_ptr s1, int start, object_ptr target);
-void Head(s1_ptr s1, int reqlen, object_ptr target);
-object Remove_elements(int start, int stop, int in_place );
+void Tail(s1_ptr s1, long start, object_ptr target);
+void Head(s1_ptr s1, long reqlen, object_ptr target);
+object Remove_elements(long start, long stop, long in_place );
 long find_from(object a, object bobj, object c);
 long e_match_from(object aobj, object bobj, object c);
 long e_match(s1_ptr a, s1_ptr b);
@@ -143,14 +143,14 @@ object Date();
 object EOpen(object filename, object mode_obj, object cleanup);
 object Command_Line();
 
-object make_atom32(unsigned c32);
+object make_atom32(unsigned long c32);
 object DoubleToInt(object d);
 
-object unary_op(int fn, object a);
+object unary_op(long fn, object a);
 
 // Binary Ops
-object binary_op_a(int fn, object a, object b);
-object binary_op(int fn, object a, object b);
+object binary_op_a(long fn, object a, object b);
+object binary_op(long fn, object a, object b);
 object x();
 object minus(long a, long b);
 object multiply(long a, long b);
